Markhor Motor Car showroom Database
This is a live project which was developed for a local CAR Agent. It is useful to the Agent to maintain the Customers details, Sales Details, Company items and services and also CAR Details
CAR Recognition System, which will cover the following modules   
The agent will enter the CAR No It will help the user to search the CAR details and the required details of the customer .The agent enters the customer details and maintain company details.
Regarding the registration module, it contains the information about newly joined User details like name of the user, password etc. CAR Details module contains the details like the Name of the CAR, send from, send to etc.  Billing and Payment Detail will contain the details like Actual Amount to be paid, EMI. Searching is having the details of the customer. Any agent will login and search the details of the customer it will show all details like customer Details, CAR name, payment type etc.